% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contour.R
\name{rasToPoly}
\alias{rasToPoly}
\title{Contour Layer}
\usage{
rasToPoly(x, nclass = 8, breaks, mask)
}
\arguments{
\item{x}{a raster or an sf contour layer.}

\item{nclass}{number of class.}

\item{breaks}{list of breaks.}

\item{mask}{a mask layer, sf object.}
}
\description{
Create a contour layer.
}
