% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{tanaka-package}
\alias{tanaka-package}
\title{Tanaka Contours Package}
\description{
The tanaka package is a simplified implementation of the Tanaka method.
Also called "relief contours method", "illuminated contour method", or
"shaded contour lines method", the Tanaka method enhances the representation
of topography on a map using shaded contour lines. North-west white contours
represent illuminated topography and south-east black contours represent
shaded topography.
}
