% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R, R/tanaka.R
\docType{package}
\name{tanaka}
\alias{tanaka}
\alias{tanaka-package}
\title{tanaka}
\usage{
tanaka(x, nclass = 8, breaks, col, mask, light = "grey95",
  dark = "grey5", shift)
}
\arguments{
\item{x}{a raster or an sf contour layer.}

\item{nclass}{number of class.}

\item{breaks}{list of breaks.}

\item{col}{a color palette (a vector of colors).}

\item{mask}{a mask layer, sf object.}

\item{light}{light shadow (NW color)}

\item{dark}{dark shadow (SE color)}

\item{shift}{size of the shadow (in map units)}
}
\description{
Tanaka pkg

Plot a tanaka map.
}
