% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contour.R
\name{tanaka_contour}
\alias{tanaka_contour}
\title{Contour Layer}
\usage{
tanaka_contour(x, nclass = 8, breaks, mask)
}
\arguments{
\item{x}{a raster or an sf contour layer.}

\item{nclass}{number of class.}

\item{breaks}{list of breaks.}

\item{mask}{a mask layer, sf object.}
}
\description{
Create a contour layer.
}
\examples{
library(tanaka)
library(raster)
library(sf)
ras <- raster(system.file("grd/elev.grd", package = "tanaka"))
iso <- tanaka_contour(x = ras)
plot(st_geometry(iso), col = c("#FBDEE1", "#F0BFC3", "#E7A1A6",
                               "#DD8287", "#D05A60", "#C03239",
                               "#721B20", "#1D0809"))
}
